backend/requirements.txt
google-genai
chromadb
python-dotenv
sentence-transformers
flask
flask-cors